Fix high cpu usage caused by fd leak
We found a problem of high CPU usage of the supervisor. This problem is
caused by continuous polling of a wrong fd in the main loop of the
supervisor. Busy polling leads to a CPU usage close to 100%. (We can
confirm this problem through the strace tool)

This issue can be reproduced by:
1. Continuously initiate arbitrary requests to supervisor through
supervisorctl
2. After the socket fd is closed, trigger the supervisor's subprocess
to rotate the log (or reopen the file)
3. If the above steps are completed within a single main loop of the
supervisor, the problem can be triggered

The reason for the problem is that supervisor relies on using
_ignore_invalid() in the main loop to close fds. This method has a flaw
that if fd is reused before _ignore_invalid() is called, then the fd may
always exist in the fd list of poll .

This commit fixes the problem. By checking the validity of the fd in the
event list in the main loop, if the fd is not in the combined_map, it is
considered to be an invalid fd and will be removed from the list.